+++ pkg/api/api.go
@@ -36,6 +36,7 @@ import (
 	"github.com/cortexproject/cortex/pkg/scheduler/schedulerpb"
 	"github.com/cortexproject/cortex/pkg/storegateway"
 	"github.com/cortexproject/cortex/pkg/storegateway/storegatewaypb"
+	"github.com/cortexproject/cortex/pkg/util/flagext"
 	"github.com/cortexproject/cortex/pkg/util/push"
 )
 
@@ -65,15 +66,15 @@ type Config struct {
 	CustomConfigHandler ConfigHandler `yaml:"-"`
 
 	// These allow and are used to configure the addition of HTTP Header fields to logs
-	LogHeaders           bool   `yaml:"LogHeaders"`
-	TargetRequestHeaders string `yaml:"TargetRequestHeaders"`
+	LogHeaders           bool                   `yaml:"LogHeaders"`
+	TargetRequestHeaders flagext.StringSliceCSV `yaml:"TargetRequestHeaders"`
 }
 
 // RegisterFlags adds the flags required to config this to the given FlagSet.
 func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&cfg.ResponseCompression, "api.response-compression-enabled", false, "Use GZIP compression for API responses. Some endpoints serve large YAML or JSON blobs which can benefit from compression.")
 	f.BoolVar(&cfg.LogHeaders, "api.LogHeaders", false, "Enable logging of header specific context information")
-	f.StringVar(&cfg.TargetRequestHeaders, "api.TargetRequestHeaders", "", "Target Headers for Request logging (if enabled) Separate multiple headers with ~, whitespace is ignored")
+	f.Var(&cfg.TargetRequestHeaders, "api.TargetRequestHeaders", "Target Headers for Request logging (if enabled) - Separate multiple headers with commas")
 	cfg.RegisterFlagsWithPrefix("", f)
 }
 
@@ -99,8 +100,7 @@ type API struct {
 	logger               log.Logger
 	sourceIPs            *middleware.SourceIPExtractor
 	indexPage            *IndexPageContent
-	HTTPHeaderMiddleware HTTPHeaderMiddleware
-	LogHeaders           bool
+	HTTPHeaderMiddleware *HTTPHeaderMiddleware
 }
 
 func New(cfg Config, serverCfg server.Config, s *server.Server, logger log.Logger) (*API, error) {
@@ -124,7 +124,6 @@ func New(cfg Config, serverCfg server.Config, s *server.Server, logger log.Logge
 		logger:         logger,
 		sourceIPs:      sourceIPs,
 		indexPage:      newIndexPageContent(),
-		LogHeaders:     cfg.LogHeaders,
 	}
 
 	// If no authentication middleware is present in the config, use the default authentication middleware.
@@ -132,7 +131,7 @@ func New(cfg Config, serverCfg server.Config, s *server.Server, logger log.Logge
 		api.AuthMiddleware = middleware.AuthenticateUser
 	}
 	if cfg.LogHeaders {
-		api.HTTPHeaderMiddleware = NewHttpHeaderMiddleware(cfg.TargetRequestHeaders)
+		api.HTTPHeaderMiddleware = &HTTPHeaderMiddleware{TargetHeaders: cfg.TargetRequestHeaders}
 	}
 
 	return api, nil
@@ -152,7 +151,7 @@ func (a *API) RegisterRoute(path string, handler http.Handler, auth bool, method
 	if a.cfg.ResponseCompression {
 		handler = gziphandler.GzipHandler(handler)
 	}
-	if a.LogHeaders {
+	if a.HTTPHeaderMiddleware != nil {
 		handler = a.HTTPHeaderMiddleware.Wrap(handler)
 	}
 
@@ -172,7 +171,7 @@ func (a *API) RegisterRoutesWithPrefix(prefix string, handler http.Handler, auth
 	if a.cfg.ResponseCompression {
 		handler = gziphandler.GzipHandler(handler)
 	}
-	if a.LogHeaders {
+	if a.HTTPHeaderMiddleware != nil {
 		handler = a.HTTPHeaderMiddleware.Wrap(handler)
 	}