+++ pkg/api/middlewares.go
@@ -3,13 +3,13 @@ package api
 import (
 	"context"
 	"net/http"
-	"regexp"
+
+	util_log "github.com/cortexproject/cortex/pkg/util/log"
+	"github.com/weaveworks/common/middleware"
 
 	"github.com/cortexproject/cortex/pkg/chunk/purger"
 	"github.com/cortexproject/cortex/pkg/querier/queryrange"
 	"github.com/cortexproject/cortex/pkg/tenant"
-	util_log "github.com/cortexproject/cortex/pkg/util/log"
-	"github.com/weaveworks/common/middleware"
 )
 
 // middleware for setting cache gen header to let consumer of response know all previous responses could be invalid due to delete operation
@@ -35,46 +35,31 @@ type HTTPHeaderMiddleware struct {
 	TargetHeaders []string
 }
 
-// NewHttpHeaderMiddleware creates a new hTTPHeaderMiddleware given a string containing headers delimited by ~ (whitespaces in headers ignored)
-func NewHttpHeaderMiddleware(unparsedTargetHeaders string) HTTPHeaderMiddleware {
-	removeWhitespaceRegex, err := regexp.Compile(`\s`)
-	if err != nil {
-		util_log.Logger.Log("WhitespaceRegexCompilationStatus", "Failed")
-		return HTTPHeaderMiddleware{TargetHeaders: nil}
-	}
-	headersWithNoWhitespace := removeWhitespaceRegex.ReplaceAllString(unparsedTargetHeaders, "")
+// InjectTargetHeadersIntoHTTPRequest injects specified HTTPHeaders into the request context
+func (h HTTPHeaderMiddleware) InjectTargetHeadersIntoHTTPRequest(r *http.Request) context.Context {
+	headerMap := make(map[string]string)
 
-	splitRegex, err2 := regexp.Compile(`~`)
-	if err2 != nil {
-		util_log.Logger.Log("SplitRegexCompilationStatus", "Failed")
-		return HTTPHeaderMiddleware{TargetHeaders: nil}
-	}
-	splitHeaders := splitRegex.Split(headersWithNoWhitespace, -1)
-	return HTTPHeaderMiddleware{TargetHeaders: splitHeaders}
-
-}
-
-// InjectRequestIdIntoHTTPRequest injects specified HTTPHeaders into the request context
-func (h HTTPHeaderMiddleware) InjectRequestIdIntoHTTPRequest(r *http.Request) context.Context {
 	// Check to make sure that Headers have not already been injected
-	testing, ok := r.Context().Value(util_log.RequestTargetsContextKey).([]string)
-	if ok && h.TargetHeaders != nil && testing[0] == h.TargetHeaders[0] {
+	testing, ok := r.Context().Value(util_log.HeaderMapContextKey).(map[string]string)
+	if ok && h.TargetHeaders != nil && testing[h.TargetHeaders[0]] == r.Header.Get(h.TargetHeaders[0]) {
 		return r.Context()
 	}
-	headerContents := make([]string, len(h.TargetHeaders))
-	for index, target := range h.TargetHeaders {
-		headerContents[index] = r.Header.Get(target)
+
+	for _, target := range h.TargetHeaders {
+		contents := r.Header.Get(target)
+		if contents != "" {
+			headerMap[target] = contents
+		}
 	}
 	ctx := r.Context()
-	ctx = context.WithValue(ctx, util_log.RequestTargetsContextKey, h.TargetHeaders)
-	ctx = context.WithValue(ctx, util_log.RequestValuesContextKey, headerContents)
+	ctx = context.WithValue(ctx, util_log.HeaderMapContextKey, headerMap)
 	return ctx
 }
 
 // Wrap implements Middleware
 func (h HTTPHeaderMiddleware) Wrap(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := h.InjectRequestIdIntoHTTPRequest(r)
+		ctx := h.InjectTargetHeadersIntoHTTPRequest(r)
 		next.ServeHTTP(w, r.WithContext(ctx))
 	})
 }