+++ pkg/util/log/wrappers.go
@@ -2,10 +2,12 @@ package log
 
 import (
 	"context"
-	"github.com/cortexproject/cortex/pkg/tenant"
+
 	"github.com/go-kit/log"
 	kitlog "github.com/go-kit/log"
 	"github.com/weaveworks/common/tracing"
+
+	"github.com/cortexproject/cortex/pkg/tenant"
 )
 
 // WithUserID returns a Logger that has information about the current user in
@@ -31,7 +33,7 @@ func WithTraceID(traceID string, l kitlog.Logger) kitlog.Logger {
 func WithContext(ctx context.Context, l kitlog.Logger) kitlog.Logger {
 	// Weaveworks uses "orgs" and "orgID" to represent Cortex users,
 	// even though the code-base generally uses `userID` to refer to the same thing.
-	l = LogHeadersFromContext(ctx, l)
+	l = HeadersFromContext(ctx, l)
 	userID, err := tenant.TenantID(ctx)
 	if err == nil {
 		l = WithUserID(userID, l)
@@ -50,13 +52,12 @@ func WithSourceIPs(sourceIPs string, l log.Logger) log.Logger {
 	return log.With(l, "sourceIPs", sourceIPs)
 }
 
-// LogHeadersFromContext enables the logging of specified HTTP Headers that have been added to a context
-func LogHeadersFromContext(ctx context.Context, l log.Logger) log.Logger {
-	targetHeaders, ok := ctx.Value(RequestTargetsContextKey).([]string)
-	headerValues, ok2 := ctx.Value(RequestValuesContextKey).([]string)
-	if ok && ok2 {
-		for index, headerName := range targetHeaders {
-			l = kitlog.With(l, headerName, headerValues[index])
+// HeadersFromContext enables the logging of specified HTTP Headers that have been added to a context
+func HeadersFromContext(ctx context.Context, l log.Logger) log.Logger {
+	testMap, ok := ctx.Value(HeaderMapContextKey).(map[string]string)
+	if ok {
+		for target, contents := range testMap {
+			l = kitlog.With(l, target, contents)
 		}
 	}
 	return l